filesystem storage of gff3 data totally encapsulated in Annotation model
 uploads/users directory renamed in uploads/private
 (uploads/public exists already)

 User model:

   - deleted all code to create / delete the uploads directory

 Annotation model:

   - now calculates automatically a filename and keeps it uptodate
   - creates and deletes uploads directories when needed
   - $GFF3_STORAGE_PATH in environment.rb defines the basis path
   - see also documentation of the class

notes: this way changing the filesystem storage to a memory- or
db-storage of the annotation will not require any change in the
code outside the Annotation class and db-table!

+++ app/models/annotation.rb
@@ -1,3 +1,40 @@
+=begin rdoc
+Represent an annotation, the content of a gff3 file, and contains 
+one or several sequence regions, described by the SequenceRegion model
+  
+* data currently kept in two positions: 
+   * metainformation ==> db-table "annotations"
+   * gff3 data       ==> filesystem
+   
+* the storage of the data in the filesystem is kept transparent to 
+  the outside of this class through following two mechanisms:
+
+(1) filename automatically calculated as:
+  
+  $GFF3_STORAGE_PATH/"public_or_private"/user_id/name
+  
+  where:
+  
+  * $GFF3_STORAGE_PATH: gives the basis path, set up in environment.rb
+    [note: it can be anywhere in the filesystem]
+  * - public annotations are kept under /public/user_id
+    - private annotations are kept under /private/user_id
+    - if the user_id directory does not exist, it is created
+  * name is a metadata saved in a column in the database
+
+  --> if the column name is changed, the file is renamed; see method name=()
+  --> if the public flag is set/unset, the file is moved; see method public=()
+  
+  (the filename is calculated in the private method "gff3_data_storage")
+
+(2) virtual attributes to access the data: 
+
+  * gff3_data 
+  * gff3_data=() 
+        
+        provide I/O access to the data in the file
+
+=end 
 class Annotation < ActiveRecord::Base
 
   ### associations ###
@@ -5,26 +42,59 @@ class Annotation < ActiveRecord::Base
   has_many :sequence_regions
   belongs_to :user
 
-  ### virtual attributes ###
-    
-  # currently the storage is in the filesystem:
-  # gff3_data is therefore implemented as virtual attribute
-  # based on gff3_data_storage, the actual column in the table
-  # which currently contains the name of the file where 
-  # the data is stored
+  ### encapsulation of the storage mechanism ###
   
+  def gff3_data_storage
+    return "#{$GFF3_STORAGE_PATH}/#{public? ? 'public' : 'private'}/#{user_id}/#{name}"
+  end
+  private :gff3_data_storage
+
+  def name=(value)
+    if new_record?
+      self[:name]=value
+    else
+      oldname = gff3_data_storage
+      self[:name]=value
+      newname = gff3_data_storage
+      File.rename(oldname, newname)      
+    end
+  end
+  
+  def public=(value)
+    if new_record?
+      self[:public]=value
+    else
+      oldname = gff3_data_storage
+      self[:public]=value
+      newname = gff3_data_storage
+      # create directory if necessary
+      user_dir = File.dirname(newname)
+      Dir.mkdir(user_dir) unless File.exists?(user_dir)
+      # move file
+      File.rename(oldname, newname)  
+      # delete old directory if empty 
+      # (if not empty it will only raise an error which is ignored -> rescue nil)
+      Dir.delete(File.dirname(oldname)) rescue nil      
+    end
+  end
+  
+  ### virtual attributes ###
+        
   def gff3_data
+    return nil unless File.exists?(gff3_data_storage)
     File.open(gff3_data_storage).read
-  end 
+  end
   def gff3_data=(data)
-    File.open(gff3_data_storage, "wb").write(data)
+    user_dir = File.dirname(gff3_data_storage)
+    Dir.mkdir(user_dir) unless File.exists?(user_dir)
+    File.open(gff3_data_storage, "wb") {|f| f.write(data)}
   end
   # see also delete_gff3_data() in the callbacks section 
 
   # a nice label for the annotation
-  def name 
+  def label 
     # (currently: filename without extension)
-    File.basename(gff3_data_storage, ".*")
+    File.basename(name, ".*")
   end
 
   ### validations ###
@@ -74,6 +144,9 @@ class Annotation < ActiveRecord::Base
   # delete the file containing the data pointed by this object
   def delete_gff3_data
     File.delete(gff3_data_storage)
+    # delete also directory if empty 
+    # (if not empty it will only raise an error which is ignored -> rescue nil)
+    Dir.delete(File.dirname(gff3_data_storage)) rescue nil
   end
 
 end
